+++ src/main.rs
@@ -1,4 +1,4 @@
-use web_dv24r::App;
+use web_sw_cor24_pcode::App;
 
 fn main() {
     console_error_panic_hook::set_once();